Rebased onto master

+++ src/main/sensors/barometer.h
@@ -33,9 +33,9 @@ typedef enum {
 typedef struct barometerConfig_s {
     uint8_t baro_hardware;                  // Barometer hardware to use
     uint8_t baro_sample_count;              // size of baro filter array
-    float baro_noise_lpf;                   // additional LPF to reduce baro noise
-    float baro_cf_vel;                      // apply Complimentary Filter to keep the calculated velocity based on baro velocity (i.e. near real velocity)
-    float baro_cf_alt;                      // apply CF to use ACC for height estimation
+    uint16_t baro_noise_lpf;                // additional LPF to reduce baro noise
+    uint16_t baro_cf_vel;                   // apply Complimentary Filter to keep the calculated velocity based on baro velocity (i.e. near real velocity)
+    uint16_t baro_cf_alt;                   // apply CF to use ACC for height estimation
 } barometerConfig_t;
 
 PG_DECLARE(barometerConfig_t, barometerConfig);
